Outsource some functionality from `square.py` into new `main.py`.
