+++ pyproject.toml
@@ -29,7 +29,7 @@ pytest-flake8 = "^1.0"
 requests-mock = "^1.7"
 
 [tool.poetry.scripts]
-square = 'square.square:main'
+square = 'square.main:main'
 
 [build-system]
 requires = ["poetry>=0.12"]
